amazon-customer-feedback-reviews-ALP-main/backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
torch==2.1.0
transformers==4.39.3
python-multipart
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
email-validator==2.1.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
tf-keras